Rename Input to TextInput
Cause that’s what it is – a TextField wrapper.
