+++ web-app/react-app/components/AuthPage/pages/ForgotPasswordPage/index.jsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import cx from 'classnames'
 
 import AuthPanel, { styles as authPanelStyles } from '../../shared/AuthPanel'
-import Input from '../../shared/Input'
+import TextInput from '../../shared/TextInput'
 import Actions from '../../shared/Actions'
 import Button from '../../shared/Button'
 import AuthStep from '../../shared/AuthStep'
@@ -55,7 +55,7 @@ class ForgotPasswordPage extends React.Component<Props> {
                     onSubmit={this.submit}
                     onFailure={this.onFailure}
                 >
-                    <Input
+                    <TextInput
                         name="email"
                         label="Email"
                         value={form.email}
